Update README.md
Hospital readmission and length-of-stay predictions provide information on how to manage hospital bed capacity and the number of required staff, especially during pandemics. We present a hybrid deep model called the Genetic Algorithm-Optimized Convolutional Neural Network (GAOCNN), with a unique preprocessing method to predict hospital readmission and the length of stay required for patients of various conditions. GAOCNN uses one-dimensional convolutional layers to predict hospital readmission and the length of stay. The parameters of the layers are optimized via a genetic algorithm. To show the performance of the proposed model in patients with various conditions, we evaluate the model under three healthcare datasets: the Diabetes 130-US hospitals dataset, the COVID-19 dataset, and the MIMIC-III dataset. The diabetes 130-US hospitals dataset has information on both readmission and the length of stay, while the COVID-19 and MIMIC-III datasets just include information on the length of stay. Experimental results show that the proposed model’s accuracy for hospital readmission was 97.2% for diabetic patients. Furthermore, the accuracy of the length-of-stay prediction was 89%, 99.4%, and 94.1% for the diabetic, COVID-19, and ICU patients, respectively. These results confirm the superiority of the proposed model compared to existing methods. Our findings offer a platform for managing the healthcare funds and resources for patients with various diseases.